requirements.txt
python-dotenv
openai
pinecone-client
pandas
openpyxl
requests
matplotlib
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pytest
